+++ lib/starcall/endpoints/endpoints.rb
@@ -1,3 +1,6 @@
 # frozen_string_literal: true
 
 require_relative 'champion_mastery_v4'
+require_relative 'champion_v3'
+require_relative 'clash_v1'
+require_relative 'summoner_v4'